New: Configurable REPL prompt
The REPL now supports the configuration of the user prompt. Usage:

:set prompt <prompt>

If the prompt starts with a letter, it is printed as-is, whereby the
sequence "%s" is expanded to the list of modules currently loaded,
and "%%" is expanded to a percent sign.

If the prompt starts with a double qoute, it is read as a string and
therefore also supports the normal escape sequences.