Revert "src/Installation.curry now uses infos from runtime/Installation.hs in order to avoid bootstrapping problems"
This reverts commit 2427fa389c6082db583276ec9ce3d7070d74db6a.

The reason for this revert is that KiCS2 can not be bootstrapped using
PAKCS since this commit. This old commit replaced the generated file
'Installation.curry' by an implementation using external functions
which use 'runtime/Installation.hs', but since no external definition
was given for PAKCS the compiler can not be translated.

The old commit mentions "bootstrapping problems" as the cause for this
commit, but the exact problems are now forgotten, unfortunately.